+++ src/index.ts
@@ -74,6 +74,7 @@ export { ReviewService } from "./internal/peek/reviews/review-service.js";
 
 export { ACTIVITY_PRODUCT_TYPE, ADD_ON_PRODUCT_TYPE, RENTAL_PRODUCT_TYPE } from "./models/peek/product.js";
 export type { Product, ProductMeetingLocation, ProductTicket } from "./models/peek/product.js";
+export type { CustomQuestion, CustomQuestionOption } from "./models/peek/custom-question.js";
 export type { AccountUser, AssignedActivity } from "./models/peek/account-user.js";
 export type {
   ResourcePool,
@@ -137,6 +138,7 @@ export type {
   CreateBookingTicket,
   CreatedBooking,
   CustomQuestionAnswer,
+  CustomQuestionAnswerInput,
   Guest,
   GuestMetadata,
   NoteMode,